feat: reorganize app
